ml/requirements.txt
# Core ML/DL
torch>=2.1.1
torchvision>=0.16.1
numpy>=1.26.2

# Computer Vision
opencv-python>=4.8.1.78
albumentations>=1.3.0
ultralytics>=8.0.0  # for YOLOv8

# Data Processing
pyyaml>=6.0.1
gdown>=5.1.0  # for downloading datasets

# MLOps
mlflow>=2.8.0

# API and Dashboard
fastapi==0.104.1
uvicorn==0.24.0
streamlit>=1.29.0
pandas>=2.1.3
plotly>=5.18.0

# Optional: GPU Support
# torch-cuda  # Uncomment if you have NVIDIA GPU 

# New dependencies
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
pydantic==2.4.2
psutil==5.9.6
tenacity==8.2.3
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
matplotlib==3.8.2
seaborn==0.13.0

# Tracing and Metrics
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlite3==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
prometheus-client==0.19.0

# Service Discovery
python-consul==1.1.0 